refactor(ai-settings): AI 設定ウィンドウをセクション別コンポーネントに分割 (#707)
AiSettingsContent.vue (1528 行) が 6 セクション + JSON5 エディタ + 共有
スタイルを 1 ファイルに抱えていたのを、window/ai-settings/ 配下の
セクションコンポーネントへ分割する。

- AiSettingsSection: 折りたたみセクションの共通クローム (ヘッダー/chip/chevron)
- AiSwitchRow: nd-toggle-switch の共通行レイアウト (8 箇所で使用)
- AiConnectionSection / AiPersonaSection / AiDataSourcesSection /
  AiHeartbeatSection: 各セクションの view + ローカルスタイル
- 親は tabs / JSON5 raw editor / import・export / reset / debounce 保存のみ
  (1528 行 → 308 行)。config は useAiConfig のシングルトンを子が直接変更し、
  親の deep watch が保存を担う従来構造を維持
- #712 の権限分離で死んでいたスタイル (toggleItem 系 / notice / warning /
  testBtn 等) を削除

見た目・挙動は変更なし。typecheck / vitest 1258 件 / vite build で確認。
